Refactored Card_Service and Module 3 Part1 Step1
