+++ templates/index.html
@@ -40,7 +40,7 @@
           </tr>
           </thead>
           <tbody>
-          {% for item in items %}
+          {% for item in view_model.items %}
           <tr>
             <td>{{item.card_name}}</td>
             <td>
